fix: system-wide audit round 1 — capture core, NaN vectors, watchdogs, image math
Multi-perspective audit (3 parallel reviewers + live black-box stress),
every finding verified before fixing; the two capture-core bugs were
also REPRODUCED live before and re-verified fixed after.

Capture core (both reproduced live):
- BLOB handler was registered once per process via a sticky bool; a
  reconnect creates a fresh IndiClient with an empty event list, so
  every capture after an indiserver bounce hung to full timeout
  forever. Handler now re-registers per client instance.
  (Repro: kill indiserver -> reconnect -> capture 30s hang; now
  completes normally.)
- Overlapping captures overwrote each other's completion source: the
  loser hung and its timeout-abort killed the winner's exposure.
  Per-device semaphore serializes exposures.
  (Repro: 2 concurrent captures both dead; now both return 200 with
  distinct frames.)

NaN crash class (same mechanism as the live-confirmed WS handshake
death): GuideHistory stored raw PHD2 distances (NaN around lost-star)
and NinaHub sends it during OnConnectedAsync -> sanitize at source;
RemoteEventBus now logs faulted SignalR sends instead of losing them
as unobserved exceptions.

Watchdogs:
- CPU killer exempted only the live-view stream; an active PHD2 guide
  loop (>15% CPU, stream off) got its guide camera SIGKILLed
  mid-session. Now exempt during Guiding/Calibrating/Looping.
- ffmpeg 'stopped draining stdin' wedge froze the push-success clock
  and was misclassified as an upstream producer stall forever (no
  recovery, frozen video, running=true). Separate attempt clock now
  distinguishes it; two strikes restart the encoder.
- Exposure-stuck watchdog collapsed to a 60s budget when a driver
  omits CCD_EXPOSURE_VALUE, aborting every legitimate >60s sub; with
  no countdown data the destructive path now waits a 15min ceiling.
- /bin/ps process handle leaked every 30s tick (no dispose/wait);
  _hungHistory mutated without the lock RecentHangs reads under.

Image math:
- FlatWizard measured ADU by reading raw bytes as UNSIGNED shorts;
  INDI writes BITPIX=16 signed with BZERO=32768, so the metric was
  top-bit-flipped garbage and the exposure search pinned flats at
  MinExposure. Now routed through FitsReader (BZERO-aware physical
  values).
- Calibration frame loop: a PNG preview failure threw away the FITS
  while still counting the frame as captured (silent data loss +
  fake success). FITS now always persists; preview failure only logs.
- Live-stack accumulator float->double: float stops absorbing
  additions past ~16.7M/pixel (~256 full-well frames), silently
  degrading exactly the deep stacks that matter.

Not yet addressed from the audits (tracked): stale-BLOB cross-delivery
hardening beyond serialization, DevicesChanged invoke unguarded in the
BLOB fast path, capture-store Id drift across restart, WebRTC peer
'disconnected' leak, safety-response abort remaining INDI-only,
sensor-staleness bounce false positives.

+++ NINA.Headless/Services/LiveStackService.cs
@@ -15,7 +15,10 @@ public class LiveStackService
     private bool _active;
     private int _width, _height;
     private List<StarDetector.Star>? _referenceStars;
-    private float[]? _sum;             // running sum of warped calibrated pixels
+    // double accumulator: float's 24-bit mantissa stops absorbing per-frame
+    // additions once a bright pixel's sum passes ~16.7M (~256 full-well frames)
+    // — exactly the deep stacks where precision matters.
+    private double[]? _sum;             // running sum of warped calibrated pixels
     private int[]? _count;             // per-pixel contribution count (edges see fewer frames)
     private int _framesAccepted;
     private int _framesRejected;
@@ -121,7 +124,7 @@ public class LiveStackService
                     _width = px.Width;
                     _height = px.Height;
                     _referenceStars = stars;
-                    _sum = new float[px.Pixels.Length];
+                    _sum = new double[px.Pixels.Length];
                     _count = new int[px.Pixels.Length];
                     for (int i = 0; i < px.Pixels.Length; i++)
                     {
@@ -178,7 +181,7 @@ public class LiveStackService
             for (int i = 0; i < _sum.Length; i++)
             {
                 var c = _count[i];
-                snapshot[i] = c > 0 ? _sum[i] / c : 0f;
+                snapshot[i] = c > 0 ? (float)(_sum[i] / c) : 0f;
             }
             w = _width;
             h = _height;
